requirements.txt
hikari-lightbulb~=2.1.1
aioredis~=2.0.0
orjson~=3.6.5
